File seperation and bug fixes on interactive mode
