+++ src/color.h
@@ -1,0 +1,15 @@
+#ifndef COLOR_H
+#define COLOR_H
+
+#define RESET  "\x1B[0m"
+#define RED  "\x1B[31m"
+#define GREEN  "\x1B[32m"
+#define YELLOW  "\x1B[33m"
+#define BLUE  "\x1B[34m"
+#define MAGENTA  "\x1B[35m"
+#define CYAN  "\x1B[36m"
+#define WHITE  "\x1B[37m"
+
+void switch_color(int col);
+
+#endif
