+++ src/xArtix.c
@@ -4,383 +4,20 @@
 #include <string.h>
 #include <time.h>
 #include <stdbool.h>
-#define RESET  "\x1B[0m"
-#define RED  "\x1B[31m"
-#define GREEN  "\x1B[32m"
-#define YELLOW  "\x1B[33m"
-#define BLUE  "\x1B[34m"
-#define MAGENTA  "\x1B[35m"
-#define CYAN  "\x1B[36m"
-#define WHITE  "\x1B[37m"
-void switch_color(int col) {
-    col = col >= 0 ? col : 0;
-    switch(col) {
-        case 0:
-            printf("%s",RESET);
-            break;
-        case 1:
-            printf("%s",RED);
-            break;
-        case 2:
-            printf("%s",GREEN);
-            break;
-        case 3:
-            printf("%s",YELLOW);
-            break;
-        case 4:
-            printf("%s",BLUE);
-            break;
-        case 5:
-            printf("%s",MAGENTA);
-            break;
-        case 6:
-            printf("%s",CYAN);
-            break;
-        case 7:
-            printf("%s",WHITE);
-            break;
-        default:
-            return;
-    }
-}
-const size_t BUF_SIZE = 2096;
+
+#include "color.h"
+#include "generator.h"
+#include "interactive.h"
+
+
 const char *version = "v1.0.1";
-typedef struct {
-    char *symb;
-    char *l_let;
-    char *u_let;
-    char *ints;
-    char *Pass;
-    int len;
-    bool prefix;
-    char *file;
-    bool failed;
-    char *pref;
-    char tmp_c;
-    char *test_symb;
-    bool test_symb_b;
-} Password;
-void i_help() {
-    printf("Available commands in xArtix interactive mode:\n");
-    printf("\t\t•\tset verbose\t\t|\t[true/false] sets the status of verbose output.\n");
-    printf("\t\t•\tset length\t\t|\t[num] sets the length for generated passwords.\n");
-    printf("\t\t•\tset output\t\t|\t[file] sets the file to output to.\n");
-    printf("\t\t•\tset unsafe\t\t|\t[true/false] sets the status for using an unsafe RNG.\n");
-    printf("\t\t•\tset extra-unicode\t|\t[true/false] sets the status for using extra unicode chars.\n");
-    printf("\t\t•\tset prompt\t\t|\t[prompt] sets the interactive prompt.\n");
-    printf("\t\t•\tgenerate\t\t|\tgenerates the prompt with select options.\n");
-}
-int pull_rand();
-void init(Password *pass, size_t BUFFER);
-void safe_return_ran(int line);
-void unsafe_return_ran(int line);
+
+
 void usage();
-void safe_set_pass(size_t len);
-void unsafe_set_pass(size_t len);
-void write_file(char *sr, char *path);
+// void write_file(char *sr, char *path);
 Password pass, *p = &pass;
-typedef struct {
-    char *prompt;
-    int len;
-    char *input;
-    bool verbose;
-    bool unsafe;
-    bool file;
-} xArtix;
-void k_parse(char *msg, xArtix *kav) {
-    char *file = (char*) malloc(256*(sizeof(char*)));
-    char *token = strtok(msg, " ");
-    if (strncmp(token, "\n", 1) == 0 || strncmp(token, "\0", 1) == 0) {
-        return;
-    }
-    else if (strncmp(token,"set", 3) == 0) {
-        token = strtok(NULL, " ");
-        if (token == NULL) {
-            fprintf(stderr, "No prompt -set- value provided.\n");
-            return;
-        }
-        if (token != NULL) {
-            if (strncmp(token, "prompt", 6) == 0) {
-                token = strtok(NULL, " ");
-                if (token == NULL) {
-                    fprintf(stderr, "No prompt provided.\n");
-                    return;
-                }
-                int len = strlen(token);
-                if (token[len - 1] == '\n') {
-                    token[len - 1] = '\0';
-                }
-                strcat(token, " ");
-                kav->prompt = token;
-            }
-            else if (strncmp(token, "prefix", 6) == 0) {
-                token = strtok(NULL, " ");
-                if (token == NULL) {
-                    fprintf(stderr, "No prefix provided.\n");
-                    return;
-                }
-                if (token[strlen(token - 1)] == '\n') {
-                    token[strlen(token - 1)] = '\0';
-                }
-                p->prefix = true;
-                p->pref = token;
-            } 
-            else if (strncmp(token, "extra-unicode", 13) == 0) {
-                token = strtok(NULL, " ");
-                if (token[strlen(token) - 1] == '\n') {
-                    token[strlen(token) - 1] = '\0';
-                }
-                if (token == NULL) {
-                    fprintf(stderr, "No setting specified.\n");
-                    return;
-                }
-                if (strncmp(token, "true", 4) == 0) {
-                    p->test_symb_b= true;
-                }
-                else if (strncmp(token, "false", 5) == 0) {
-                    p->test_symb_b = false;
-                }    
-                else {
-                    fprintf(stderr, "Option: \"%s\" is not valid.\n",token);
-                }
-            }
-            else if (strncmp(token, "unsafe", 6) == 0) {
-                token = strtok(NULL, " ");
-                if (token == NULL) {
-                    fprintf(stderr, "No setting specified.\n");
-                    return;
-                }
-                if (token[strlen(token) - 1] == '\n') {
-                    token[strlen(token) - 1] = '\0';
-                }
-                if (strncmp(token, "true", 4) == 0) {
-                    kav->unsafe = true;
-                }
-                else if (strncmp(token, "false", 5) == 0) {
-                    kav->unsafe = false;
-                }    
-                else {
-                    fprintf(stderr, "Option: \"%s\" is not valid.\n",token);
-                }
-            }
-            else if (strncmp(token, "color", 5) == 0) {
-                token = strtok(NULL, " ");
-                if (token == NULL) {
-                    fprintf(stderr, "No color specified.\n");
-                    return;
-                }
-                if (token[strlen(token) - 1] == '\n') {
-                    token[strlen(token) - 1] = '\0';
-                }
-                if (strncmp(token, "reset", 5) == 0) {
-                    switch_color(0);
-                } else if (strncmp(token, "red", 3) == 0) {
-                    switch_color(1);
-                } else if (strncmp(token, "green", 5) == 0) {
-                    switch_color(2);
-                } else if (strncmp(token, "yellow", 6) == 0) {
-                    switch_color(3);
-                } else if (strncmp(token, "blue", 4) == 0) {
-                    switch_color(4);
-                } else if (strncmp(token, "magenta", 7) == 0) {
-                    switch_color(5);
-                } else if (strncmp(token, "cyan", 4) == 0) {
-                    switch_color(6);
-                } else if (strncmp(token, "white", 5) == 0) {
-                    switch_color(7);
-                } else {
-                    fprintf(stderr, "Color: \"%s\" is not recognized.\n",token);
-                }
-            }
-            else if (strncmp(token, "length", 6) == 0) {
-                token = strtok(NULL, " ");
-                if (token == NULL) {
-                    fprintf(stderr, "No length specified.\n");
-                    return;
-                }
-                if (token[strlen(token - 1)] == '\n') {
-                    token[strlen(token - 1)] = '\0';
-                }
-                if (atoi(token) > 2096) {
-                    fprintf(stderr, "Length cannot be over 2096.\n");
-                    return;
-                }
-                kav->len = atoi(token);
-            }
-            else if (strncmp(token, "output", 6) == 0) {
-                token = strtok(NULL, " ");
-                if (token == NULL) {
-                    fprintf(stderr, "No output file specified.\n");
-                }
-                if (token[strlen(token) - 1] == '\n') {
-                    token[strlen(token) - 1] = '\0';
-                }
-                file = token;
-                kav->file = true;
-            }
-            else if (strncmp(token, "verbose", 7) == 0) {
-                token = strtok(NULL, " ");
-                if (token == NULL) {
-                    fprintf(stderr, "No setting specified.\n");
-                    return;
-                }
-                if (token[strlen(token) - 1] == '\n') {
-                    token[strlen(token) - 1] = '\0';
-                }
-                if (strncmp(token, "true", 4) == 0) {
-                    kav->verbose = true;
-                }
-                else if (strncmp(token, "false", 5) == 0) {
-                    kav->verbose = false;
-                }    
-                else {
-                    fprintf(stderr, "Option: \"%s\" is not valid.\n",token);
-                }
-            }
-            else {
-                fprintf(stderr, "Setting not recognized.\n");
-            }
-        }
-    }
-    else if (strncmp(token, "help", 4) == 0) {
-        i_help();
-    }
-    else if (strncmp(token, "generate", 8) == 0) {
-        if (!kav->unsafe) {
-            safe_set_pass(kav->len);
-            if (kav->verbose) {
-                if (kav->file) {
-                    write_file(p->Pass, file);
-                    printf("Password: %s\nWith length: %d written to file: %s\n",p->Pass, kav->len, file);
-                }
-                else {
-                    printf("Password: %s\nWith length: %d\n",p->Pass, kav->len);
-                }
-            }
-            else {
-                printf("%s\n",p->Pass);
-            }
-        } else {
-            unsafe_set_pass(kav->len);
-            if (kav->verbose) {
-                if (kav->file) {
-                    write_file(p->Pass, file);
-                    printf("Unsafe password: %s\nWith lengh: %d written to file: %s\n",p->Pass, kav->len, file); 
-                }
-                else {
-                    printf("Unsafe password: %s\nWith lengh: %d\n",p->Pass, kav->len);
-                }
-            }
-            else {
-                printf("%s\n",p->Pass);
-            }
-        }
-        return;
-    }
-    else if (strncmp(token, "exit", 4) == 0) {
-        return;
-    }
-    else if (strncmp(token, "clear", 5) == 0) {
-        #define ESC    "\x1b"
-        printf(ESC"[2J"ESC"[?6h");
-    }
-    else {
-        printf("Command not recognized!\n");
-    }
-    return;
-    free(file);
-}
-void k_init(xArtix *kav) {
-    kav->prompt = "> ";
-    kav->input = ";";
-}
-void k_ctl(xArtix *kav) { 
-    char buff[256];
-    k_init(kav);
-    while((strncmp(kav->input, "exit",4))) {
-        printf("%s",kav->prompt);
-        fgets(buff, 256, stdin);
-        kav->input = buff;
-        k_parse(kav->input, kav);
-    }
-}
-#define ARR_S(arr) (sizeof(arr) / sizeof((arr)[0])) 
-int pull_rand() {
-    FILE * file = fopen("/dev/random", "r");            
-    unsigned char buff[10];
-    if (!file) {
-        perror("fopen");
-        return EXIT_FAILURE;
-    }
-    size_t retv = fread(buff, sizeof(*buff), ARR_S(buff), file); /* I have to add some error checking */
-    if (retv != ARR_S(buff)) {
-        perror("fread(): failed.");
-        return EXIT_FAILURE;
-    }
-    int temp = buff[0];
-    fclose(file);
-    return temp;
-}
-void init(Password *pass, size_t BUFFER) {
-    pass->symb = "!@#$%^&*()-=+_][{}";
-    pass->l_let = "qwertyuiopasdfghjklzxcvbnm";
-    pass->u_let = "QWERTYUIOPASDFGHJKLZXCVBNM";
-    pass->ints = "1234567890";
-    pass->test_symb = "™€‰—®©☺⚛";
-    pass->Pass = (char *) malloc(BUFFER * sizeof(char));
-}
-void safe_return_ran(int line) {
-    if (line >=6) {
-        p->failed = true;
-    }else {
-        switch(line) {
-            case 1:
-                p->tmp_c = p->symb[pull_rand() % strlen(p->symb)];
-                break;
-            case 2:
-                p->tmp_c = p->l_let[pull_rand() % strlen(p->l_let)];
-                break;
-            case 3:
-                p->tmp_c = p->u_let[pull_rand() % strlen(p->u_let)];
-                break;
-            case 4:
-                p->tmp_c = p->ints[pull_rand() % strlen(p->ints)];
-                break;
-            case 5:
-                p->tmp_c = p->test_symb[pull_rand() % strlen(p->test_symb)];
-                break;
-            default:
-                p->failed = true;
-                break;
-        }
-    }
-}
-void unsafe_return_ran(int line) {
-    if (line >=6) {
-        p->failed = true;
-    }else {
-        switch(line) {
-            case 1:
-                p->tmp_c = p->symb[rand() % strlen(p->symb)];
-                break;
-            case 2:
-                p->tmp_c = p->l_let[rand() % strlen(p->l_let)];
-                break;
-            case 3:
-                p->tmp_c = p->u_let[rand() % strlen(p->u_let)];
-                break;
-            case 4:
-                p->tmp_c = p->ints[rand() % strlen(p->ints)];
-                break;
-            case 5:
-                p->tmp_c = p->test_symb[rand() % strlen(p->test_symb)];
-                break;
-            default:
-                p->failed = true;
-                break;
-        }
-    }
-}
+
+
 void usage() {
     printf("Program: xArtix | version: %s\n", version);
     printf("Usage: xArtix -[hloveFpci]\n");
@@ -394,65 +31,7 @@ void usage() {
     printf("\t-c\t--color COLOR\t\t|\tSets a color for output.\n");
     printf("\t-i\t--interactive\t\t|\tEnters an interactive mode. (IN DEVELOPMENT)\n");
 }
-void safe_set_pass(size_t len) {
-    LOOP:
-    if (p->prefix) {
-        for(size_t i = 0; i < len + strlen(p->pref); i++) {
-            if (i >= strlen(p->pref)) {
-                if (p->test_symb_b) {
-                    safe_return_ran(pull_rand() % 5 + 1);
-                } else {
-                    safe_return_ran(pull_rand() % 4 + 1);
-                }
-                p->Pass[i] = p->tmp_c;
-                if (p->failed) {
-                    return;
-                }
-            }
-            else {
-                p->Pass[i] = p->pref[i];
-            }
-        }
-        if (strlen(p->Pass) != len + strlen(p->pref)) {
-            goto LOOP;
-            printf("not");
-        }
-    } 
-    else {
-        for(size_t i = 0; i < len; i++) {
-            if (p->test_symb_b) {
-                safe_return_ran(pull_rand() % 5 + 1);
-            } else {
-                safe_return_ran(pull_rand() % 4 + 1);
-            }
-            p->Pass[i] = p->tmp_c;
-            if (p->failed) {
-                return;
-            }
-        }
-        if (strlen(p->Pass) != len) {
-            goto LOOP;
-        }
-    }
-}
-void unsafe_set_pass(size_t len) {
-    LOOP:
-    srand(time(NULL));
-    for(size_t i = 0; i < len; i++) {
-        if (p->test_symb_b) {
-            unsafe_return_ran(rand() % 5 + 1);
-        } else {
-            unsafe_return_ran(rand() % 4 + 1);
-        }
-        p->Pass[i] = p->tmp_c;
-        if(p->failed) {
-            return;
-        }
-    }
-    if (strlen(p->Pass) != len) {
-        goto LOOP;
-    }
-}
+
 struct option long_options[] = {
     { "length",             required_argument,  0,      'l' },
     { "help",               no_argument,        0,      'h' },
@@ -465,11 +44,11 @@ struct option long_options[] = {
     { "prefix",             required_argument,  0,      'p' },
     { 0, 0, 0, 0 }
 };
-void write_file(char *str, char *path) {
-    FILE *ptr = fopen(path, "w");
-    fputs(str, ptr);
-    fclose(ptr);
-}
+// void write_file(char *str, char *path) {
+//     FILE *ptr = fopen(path, "w");
+//     fputs(str, ptr);
+//     fclose(ptr);
+// }
 int main(int argc, char **argv) {
     int c, option_index = 0;
     bool unsafe = false, verbose = false, made_pass = false, tmp = false, commence = false;
@@ -514,10 +93,8 @@ int main(int argc, char **argv) {
                 }
                 break;
             case 'i': ;
-                xArtix *kav = malloc(256 * sizeof(xArtix));
-                init(p,BUF_SIZE);
-                k_ctl(kav);
-                free(kav);
+                // Launch interactive mode
+                interactive_mode(p);
                 return 0;
                 break;
             case 'e':
@@ -552,8 +129,8 @@ int main(int argc, char **argv) {
     }
     if (!unsafe) {
         if (commence) {
-            init(p,BUF_SIZE);
-            safe_set_pass(p->len);
+            init(p);
+            safe_set_pass(p);
             if (verbose) {
                 printf("Password: %s\nWith Length: %zu\n",p->Pass, strlen(p->Pass));
             }
@@ -574,8 +151,8 @@ int main(int argc, char **argv) {
         }
     } else {
         if (commence) {
-            init(p,BUF_SIZE);
-            unsafe_set_pass(p->len);
+            init(p);
+            unsafe_set_pass(p);
             if (verbose) {
                 printf("Unsafe password: %s\nWith Length: %zu\n",p->Pass, strlen(p->Pass));
             }
@@ -602,5 +179,8 @@ int main(int argc, char **argv) {
     switch_color(0);
     if (p->Pass) {
         free(p->Pass);
+    }
+    if (p->file) {
+        free(p->file);
     }
 }
